projet-Data-eng/pipelines-main/pipelines-main/requirements.txt
apache-airflow
pendulum
requests
minio